added backup condition

+++ install.sh
@@ -6,7 +6,7 @@ set -e
 cd "$(dirname "$0")"
 
 #Caso nao tenha Stow instalado, fecha o programa
-if ! command -v stow >/dev/null 2>&1; then
+if ! command -v stow -v >/dev/null 2>&1; then
   echo "Erro: GNU Stow não está instalado."
   exit 1  
 fi 
@@ -16,10 +16,16 @@ fi
 function install_nvim {
   if [[ -d "$CONFIG_PATH/nvim" ]]; then
     echo "Há uma configuração ativa. Vai ser criado "nvim-backup" com as configurações antigas."
+    
+    if [[ -e "$CONFIG_PATH/nvim-backup" ]]; then
+      echo "Já existe um backup em $CONFIG_PATH/nvim-backup"
+      exit 1
+    fi
+    
     mv "$CONFIG_PATH/nvim" "$CONFIG_PATH/nvim-backup" && echo "Backup criado!" 
-    stow --target="$HOME" nvim && echo "Configuração instalada com sucesso!"
+    stow -v --target="$HOME" nvim && echo "Configuração instalada com sucesso!"
   else 
-   stow --target="$HOME" nvim && echo "Configuração instalada com sucesso!"
+   stow -v --target="$HOME" nvim && echo "Configuração instalada com sucesso!"
   fi 
 }
 
@@ -27,40 +33,65 @@ function install_nvim {
 function install_hypr {
   if [[ -d "$CONFIG_PATH/hypr" ]]; then
     echo "Há uma configuração ativa. Vai ser criado "hypr-backup" com as configurações antigas."
+   
+    if [[ -e "$CONFIG_PATH/hypr-backup" ]]; then
+      echo "Já existe um backup em $CONFIG_PATH/hypr-backup"
+      exit 1
+    fi 
+
     mv "$CONFIG_PATH/hypr" "$CONFIG_PATH/hypr-backup" && echo "Backup criado!" 
-    stow --target="$HOME" hypr && echo "Configuração instalada com sucesso!"
+    stow -v --target="$HOME" hypr && echo "Configuração instalada com sucesso!"
   else 
-   stow --target="$HOME" hypr && echo "Configuração instalada com sucesso!"
+   stow -v --target="$HOME" hypr && echo "Configuração instalada com sucesso!"
   fi 
 }
 
 function install_waybar {
   if [[ -d "$CONFIG_PATH/waybar" ]]; then
     echo "Há uma configuração ativa. Vai ser criado "waybar-backup" com as configurações antigas."
+    
+    
+    if [[ -e "$CONFIG_PATH/waybar-backup" ]]; then
+      echo "Já existe um backup em $CONFIG_PATH/waybar-backup"
+      exit 1
+    fi
+
     mv "$CONFIG_PATH/waybar" "$CONFIG_PATH/waybar-backup" && echo "Backup criado!" 
-    stow --target="$HOME" waybar && echo "Configuração instalada com sucesso!"
+    stow -v --target="$HOME" waybar && echo "Configuração instalada com sucesso!"
   else 
-   stow --target="$HOME" waybar && echo "Configuração instalada com sucesso!"
+   stow -v --target="$HOME" waybar && echo "Configuração instalada com sucesso!"
   fi 
 }
 
 function install_wofi {
   if [[ -d "$CONFIG_PATH/wofi" ]]; then
     echo "Há uma configuração ativa. Vai ser criado "wofi-backup" com as configurações antigas."
+    
+    if [[ -e "$CONFIG_PATH/wofi-backup" ]]; then
+      echo "Já existe um backup em $CONFIG_PATH/wofi-backup"
+      exit 1
+    fi
+
     mv "$CONFIG_PATH/wofi" "$CONFIG_PATH/wofi-backup" && echo "Backup criado!" 
-    stow --target="$HOME" wofi && echo "Configuração instalada com sucesso!"
+    stow -v --target="$HOME" wofi && echo "Configuração instalada com sucesso!"
   else 
-   stow --target="$HOME" wofi && echo "Configuração instalada com sucesso!"
+   stow -v --target="$HOME" wofi && echo "Configuração instalada com sucesso!"
   fi 
 }
 
 function install_kitty {
   if [[ -d "$CONFIG_PATH/kitty" ]]; then
     echo "Há uma configuração ativa. Vai ser criado "kitty-backup" com as configurações antigas."
+    
+    if [[ -e "$CONFIG_PATH/kitty-backup" ]]; then
+      echo "Já existe um backup em $CONFIG_PATH/kitty-backup"
+      exit 1
+    fi
+    
     mv "$CONFIG_PATH/kitty" "$CONFIG_PATH/kitty-backup" && echo "Backup criado!" 
-    stow --target="$HOME" kitty && echo "Configuração instalada com sucesso!"
+    stow -v --target="$HOME" kitty && echo "Configuração instalada com sucesso!"
   else 
-   stow --target="$HOME" kitty && echo "Configuração instalada com sucesso!"
+   stow -v --target="$HOME" kitty && echo "Configuração instalada com sucesso!"
   fi 
 }
